1.5 Hid fields, authentication msg on top

+++ functions/Ixwp_Social_Comments_Plugin.php
@@ -52,6 +52,7 @@ class Ixwp_Social_Comments_Plugin
             add_action('wp_enqueue_scripts', array($this, 'wp_enqueue_scripts'));
             add_filter('wp_get_current_commenter', array($this, 'wp_get_current_commenter'));
             add_filter('comments_open', array($this, 'comments_open'), 10, 2);
+            add_action('comment_form_top', array($this, 'comment_form_top'));
             add_action('comment_form_after', array($this, 'comment_form_after'));
             add_action('preprocess_comment', array($this, 'preprocess_comment'));
         }
@@ -158,44 +159,55 @@ class Ixwp_Social_Comments_Plugin
         }
     }
 
+  function comment_form_top()
+    {
+        global $ixwp_sc_post_protected;
+        $commenter = $this->sc_get_current_commenter();
+
+        // If comments Enabled and logged in 
+        if($ixwp_sc_post_protected && isset($commenter)){
+            
+        
+            if (array_key_exists('post_id', $_REQUEST)) {
+                $post_id = $_REQUEST['post_id'];
+            } else {
+                $post_id = get_the_ID();
+            }
+            $referrer = esc_attr(get_permalink($post_id));
+            $logout_url = admin_url('admin-ajax.php?action=ixwp-sc-logout&amp;_wp_http_referer=' . $referrer . '#ixwp-social-comment-wrapper');
+            echo '<div id="ixwp-social-comment-wrapper">';
+            echo '<p class="ixwp-social-comment-logout">';
+            echo __('You are authenticated as '.$commenter['display_name'].' in '.$commenter['network'].'. ', IXWP_SOCIAL_COMMENTS_NAME);
+            echo '<a href="' . $logout_url . '">' . __('Logout', IXWP_SOCIAL_COMMENTS_NAME) . '</a>';
+            echo '</p>';
+            echo '</div>';
+
+            // FK Hide completed comment fields
+            echo '<style>.comment-form-author, .comment-form-email, .comment-form-url {display:none;}</style>';
+        
+        }
+    }
+
     function comment_form_after()
     {
         global $ixwp_sc_post_protected;
+        $commenter = $this->sc_get_current_commenter();
 
-        // If comments Enabled:
-        if($ixwp_sc_post_protected){
+        // If comments Enabled and not logged in
+        if($ixwp_sc_post_protected && !isset($commenter)){
             
-            // If user logged in show logout button
-            $commenter = $this->sc_get_current_commenter();
-            if (isset($commenter)) {
-                if (array_key_exists('post_id', $_REQUEST)) {
-                    $post_id = $_REQUEST['post_id'];
-                } else {
-                    $post_id = get_the_ID();
+            echo '<div id="ixwp-social-comment-wrapper">';
+            echo '<div class="ixwp-social-comment-buttons">';
+            $tabs = $this->tabs;
+            foreach ($tabs as $id => $instance) {
+                if ($instance instanceof Ixwp_Social_Network_Authenticator) {
+                    echo $instance->get_auth_button();
                 }
-                $referrer = esc_attr(get_permalink($post_id));
-                $logout_url = admin_url('admin-ajax.php?action=ixwp-sc-logout&amp;_wp_http_referer=' . $referrer . '#ixwp-social-comment-wrapper');
-                echo '<div id="ixwp-social-comment-wrapper">';
-                echo '<p class="ixwp-social-comment-logout">';
-                echo __('You are authenticated as '.$commenter['display_name'].' in '.$commenter['network'].'. ', IXWP_SOCIAL_COMMENTS_NAME);
-                echo '<a href="' . $logout_url . '">' . __('Logout', IXWP_SOCIAL_COMMENTS_NAME) . '</a>';
-                echo '</p>';
-                echo '</div>';
             }
-            // Else show Social Buttons
-            else {
-                echo '<div id="ixwp-social-comment-wrapper">';
-                echo '<div class="ixwp-social-comment-buttons">';
-                $tabs = $this->tabs;
-                foreach ($tabs as $id => $instance) {
-                    if ($instance instanceof Ixwp_Social_Network_Authenticator) {
-                        echo $instance->get_auth_button();
-                    }
-                }
-                echo '</div>';
-                echo '<p class="ixwp-social-comment-wait" style="display: none;"><i class="fa fa-spinner fa-spin"></i> ' . __('Please wait while you are being authenticated...', IXWP_SOCIAL_COMMENTS_NAME) . '</p>';
-                echo '</div>';
-            }
+            echo '</div>';
+            echo '<p class="ixwp-social-comment-wait" style="display: none;"><i class="fa fa-spinner fa-spin"></i> ' . __('Please wait while you are being authenticated...', IXWP_SOCIAL_COMMENTS_NAME) . '</p>';
+            echo '</div>';
+       
         }
     }
